Extract rolling metrics router
