+++ tests/support/app_runtime.py
@@ -6,6 +6,7 @@ from typing import Any, Iterator
 import app.main as main_module
 import app.routers.drawdown as drawdown_module
 import app.routers.risk_calculation as risk_calculation_module
+import app.routers.rolling as rolling_module
 from app.main import app
 
 
@@ -29,7 +30,9 @@ def override_app_runtime(
     original_risk_performance_class: Any = getattr(
         risk_calculation_module, "LotusPerformanceClient"
     )
+    original_rolling_performance_class: Any = getattr(rolling_module, "LotusPerformanceClient")
     original_core_class: Any = getattr(main_module, "LotusCoreClient")
+    original_rolling_core_class: Any = getattr(rolling_module, "LotusCoreClient")
 
     try:
         if lotus_performance_client is not _UNSET:
@@ -44,8 +47,10 @@ def override_app_runtime(
                 "LotusPerformanceClient",
                 lotus_performance_class,
             )
+            setattr(rolling_module, "LotusPerformanceClient", lotus_performance_class)
         if lotus_core_class is not _UNSET:
             setattr(main_module, "LotusCoreClient", lotus_core_class)
+            setattr(rolling_module, "LotusCoreClient", lotus_core_class)
         if dependency_statuses is not _UNSET:
             app.state.dependency_statuses = dependency_statuses
         yield
@@ -60,4 +65,6 @@ def override_app_runtime(
             "LotusPerformanceClient",
             original_risk_performance_class,
         )
+        setattr(rolling_module, "LotusPerformanceClient", original_rolling_performance_class)
         setattr(main_module, "LotusCoreClient", original_core_class)
+        setattr(rolling_module, "LotusCoreClient", original_rolling_core_class)
